Okay, so THIS is the last 1.8 commit. No really. Includes accessibility and a help funciton.
IMPORTANT, THIS IS THE LAST BUILD THAT WILL SUPPORT NAGIOS XI PRE-5.5.X!!!

The changes in the API in 5.5.X are huge, the JSON returns were cleaned up and I don't feel like doing that much checking. So this is it for pre-5.5.X. The next release will be in the 2.0 version, and only support 5.5.X and later.